+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## 1.54.0
 
+* Deprecate selectors with leading or trailing combinators, or with multiple
+  combinators in a row. If they're included in style rules after nesting is
+  resolved Sass will now produce a deprecation warning and, in most cases, omit
+  the selector. Leading and trailing combinators can still be freely used for
+  nesting purposes.
+
+  See https://sass-lang.com/d/bogus-combinators for more details.
+
 ### JS API
 
 * Add a `charset` option that controls whether or not Sass emits a